Move base interfaces of the domain into a boundary package
